Use OrderedDict everywhere in charstree
The problem is that we perform random.choice(list(values)) in
various places. In Python 3.3+ this adds an additional element of
non-determinism because of hash randomization.

The best solution is probably don't do that then, but for now just
replacing all the dictionaries with OrderedDict works by making it
stable (because the build order is already stable)

I have very few good idea about how to test this case. :(